python_edge/requirements.txt
deepface
tf-keras
opencv-python
mediapipe
python-dotenv
supabase
pyttsx3
numpy
pandas